+++ green/src/main/java/com/blockstream/green/ui/receive/ReceiveFragment.kt
@@ -59,11 +59,7 @@ class ReceiveFragment : WalletFragment<ReceiveFragmentBinding>(
         }
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        if(isFinishingGuard) return
-
+    override fun onViewCreatedGuarded(view: View, savedInstanceState: Bundle?) {
         binding.vm = viewModel
 
         binding.address.setOnClickListener {