+++ green/src/main/java/com/blockstream/green/ui/recovery/RecoveryCheckFragment.kt
@@ -43,11 +43,7 @@ class RecoveryCheckFragment : WalletFragment<RecoveryCheckFragmentBinding>(
         return args.wallet != null
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        if(isFinishingGuard) return
-
+    override fun onViewCreatedGuarded(view: View, savedInstanceState: Bundle?) {
         binding.vm = viewModel
         binding.isDevelopmentFlavor = requireContext().isDevelopmentFlavor()