+++ green/src/main/java/com/blockstream/green/ui/settings/WalletSettingsFragment.kt
@@ -96,9 +96,7 @@ class WalletSettingsFragment :
             updateSharedPreferencesSummaries()
         }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
+    override fun onViewCreatedGuarded(view: View, savedInstanceState: Bundle?) {
         getNavigationResult<Boolean>()?.observe(viewLifecycleOwner) {
             it?.let {
                 viewModel.updateTwoFactorConfig()