+++ green/src/main/java/com/blockstream/green/ui/wallet/AddAccountFragment.kt
@@ -33,11 +33,7 @@ class AddAccountFragment : WalletFragment<AddAccountFragmentBinding>(
         AddAccountViewModel.provideFactory(viewModelFactory, wallet, args.accountType)
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        if(isFinishingGuard) return
-
+    override fun onViewCreatedGuarded(view: View, savedInstanceState: Bundle?) {
         binding.vm = viewModel
 
         binding.buttonContinue.setOnClickListener {